+++ pixmap.c
@@ -8,7 +8,7 @@ void* pixmap_free( struct pixmap* img ) {
 		fprintf( stderr, "pixmap warning: pixmap pointer == NULL\n" );
 	} else {
 		if ( NULL != img->pixels2 ) {
-			for ( long h = 0; h < img->height; h++ ) {
+			for ( int h = 0; h < img->height; h++ ) {
 				free( img->pixels2[ h ] );
 			}
 		}
@@ -26,9 +26,9 @@ void* pixmap_alloc_error( struct pixmap* img ) {
 	return NULL;
 }
 
-struct pixmap* pixmap_alloc( long width, long height ) {
+struct pixmap* pixmap_alloc( int width, int height ) {
 	if ( width <= 0 || height <= 0 ) {
-		fprintf( stderr, "pixmap error: width = %ld, height = %ld\n", width, height );
+		fprintf( stderr, "pixmap error: width = %d, height = %d\n", width, height );
 		return NULL;
 	}
 
@@ -49,18 +49,18 @@ struct pixmap* pixmap_alloc( long width, long height ) {
 	img->pixels2 = calloc( height, sizeof( unsigned char** ) );
 	if ( NULL == img->pixels2 ) return pixmap_alloc_error( img );
 
-	for ( long h = 0; h < height; h++ ) img->pixels2[ h ] = NULL; // for clean freeing
-	for ( long h = 0; h < height; h++ ) {
+	for ( int h = 0; h < height; h++ ) img->pixels2[ h ] = NULL; // for clean freeing
+	for ( int h = 0; h < height; h++ ) {
 		img->pixels2[ h ] = calloc( width, sizeof( unsigned char* ) );
 		if ( NULL == img->pixels2[ h ] ) return pixmap_alloc_error( img );
 	}
 
 	// calculate pointers
-	for ( long p = 0; p < width*height; p++ ) {
+	for ( int p = 0; p < width*height; p++ ) {
 		img->pixels[ p ] = img->bytes + PIXMAP_COLORS*p;
 	}
-	for ( long h = 0; h < height; h++ ) {
-		for ( long w = 0; w < width; w++ ) {
+	for ( int h = 0; h < height; h++ ) {
+		for ( int w = 0; w < width; w++ ) {
 			img->pixels2[ h ][ w ] = img->bytes + PIXMAP_COLORS*h*width + PIXMAP_COLORS*w;
 		}
 	}
