+++ pixmap_jpg.c
@@ -63,7 +63,7 @@ int pixmap_write_jpg( const struct pixmap* img, const char* filename ) {
 		err = 1;
 		goto free_jpg;
 	}
-	for ( long h = 0; h < img->height; h++ ) {
+	for ( int h = 0; h < img->height; h++ ) {
 		rows[ h ] = img->bytes + PIXMAP_COLORS*h*img->width;
 	}
 	if ( img->height != jpeg_write_scanlines( jpg, rows, img->height ) ) {
